[Homework_21] - Добавление вызова методов удаления, добавления и получения контакта через viewModel, изменение типа данных для номера телефона с Int на Long

+++ task_21/serg/app/src/main/java/com/weather/task7_3notebook/model/Contact.kt
@@ -3,6 +3,6 @@ package com.weather.task7_3notebook.model
 data class Contact(
     val name: String,
     val lastName: String,
-    val number: Int,
+    val number: Long,
     val city: City?
 )
